% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paintr_ui.R
\name{argsPositionUI}
\alias{argsPositionUI}
\title{ui component for setting \code{position} in \verb{geom_<chart>}}
\usage{
argsPositionUI(id, ui_id = "argsPosition")
}
\arguments{
\item{id}{An ID string that corresponds with the ID used to call the module's UI function.}

\item{ui_id}{id of the ui element}
}
\value{
a list with two elements
\itemize{
\item ui: the ui element of this piece
\item id: the id of this ui element
}
}
\description{
ui component for setting \code{position} in \verb{geom_<chart>}
}
