% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paintr-utils.R
\name{break_sum}
\alias{break_sum}
\title{Recursively Split Plot Expressions}
\usage{
break_sum(x)
}
\arguments{
\item{x}{An expression or call.}
}
\value{
A split representation of the expression.
}
\description{
Recursively Split Plot Expressions
}
\keyword{internal}
