% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paintr_control.R
\name{callFuncUI}
\alias{callFuncUI}
\title{Calls the selected ui functions by the name of the key}
\usage{
callFuncUI(name, defaultArgs, scope, extraFunc = NULL, extraFuncArgs = NULL)
}
\arguments{
\item{name}{name of the key}

\item{defaultArgs}{some default arguments passed to the ui functions}

\item{scope}{\code{scope} can be one of \code{mapping}, \code{geom_args}, or \code{plot_settings}}

\item{extraFunc}{optional. A named list of extra functions provided by the user.
For example \code{list(param1 = my_func1, param2 = my_func2)}}

\item{extraFuncArgs}{optional. A list of function arguments provided by the user.
Function arguments of one function should be formed in a list as one element of \code{extraFuncArgs}
For example \code{list(param1 = list(my_func1_arg1, my_func1_arg2), param2 = list(my_func2_arg1, my_func2_arg2))}}
}
\value{
function call of the selected function
}
\description{
Calls the selected ui functions by the name of the key
}
