% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paintr-utils.R
\name{check_remove_null}
\alias{check_remove_null}
\title{Drop \code{NULL} Elements from a List}
\usage{
check_remove_null(x)
}
\arguments{
\item{x}{A list or \code{NULL}.}
}
\value{
A list without \code{NULL} values, or \code{NULL}.
}
\description{
Drop \code{NULL} Elements from a List
}
\keyword{internal}
