% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geomGenerator.R
\name{colorGenerator}
\alias{colorGenerator}
\title{Title}
\usage{
colorGenerator(id, dataColor, fillID, scaleColorID, colorPalette = "RdYlBu")
}
\arguments{
\item{colorPalette}{}
}
\value{

}
\description{
Title
}
