% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paintr_control.R
\name{controlUI}
\alias{controlUI}
\title{get all ui elements}
\usage{
controlUI(
  id,
  defaultArgs,
  mapping,
  geom_args = NULL,
  plot_settings = NULL,
  extra_uiFunc = NULL,
  extra_uiFuncArgs = NULL
)
}
\arguments{
\item{id}{An ID string that corresponds with the ID used for all component of this \code{paintr_obj}}

\item{defaultArgs}{some default arguments passed to the ui functions}

\item{mapping}{keys of mapping}

\item{geom_args}{keys of geom_args}

\item{plot_settings}{keys of plot_settings}

\item{extra_uiFunc}{optional. A named list of extra functions provided by the user.
For example \code{list(param1 = my_func1, param2 = my_func2)}}

\item{extra_uiFuncArgs}{optional. A list of function arguments provided by the user.
Function arguments of one function should be formed in a list as one element of \code{extraFuncArgs}
For example \code{list(param1 = list(my_func1_arg1, my_func1_arg2), param2 = list(my_func2_arg1, my_func2_arg2))}}
}
\value{
a list of ui and id
}
\description{
get all ui elements
}
