% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paintr-utils.R
\name{detect_keywords}
\alias{detect_keywords}
\title{Detect Paintr Placeholder Keywords}
\usage{
detect_keywords(.x)
}
\arguments{
\item{.x}{An expression object.}
}
\value{
The placeholder keyword string or \code{NULL}.
}
\description{
Detect Paintr Placeholder Keywords
}
\keyword{internal}
