% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paintr-utils.R
\name{encode_id}
\alias{encode_id}
\title{Build a Stable Placeholder Id}
\usage{
encode_id(index_path, func_name)
}
\arguments{
\item{index_path}{An index path.}

\item{func_name}{A call name.}
}
\value{
A single encoded id string.
}
\description{
Build a Stable Placeholder Id
}
\keyword{internal}
