% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paintr-runtime.R
\name{expr_apply_checkbox_result}
\alias{expr_apply_checkbox_result}
\title{Apply a Layer Checkbox Result}
\usage{
expr_apply_checkbox_result(expr, nn, input)
}
\arguments{
\item{expr}{A layer expression.}

\item{nn}{The layer name.}

\item{input}{A Shiny input-like object.}
}
\value{
The layer expression or \code{NULL}.
}
\description{
Apply a Layer Checkbox Result
}
\keyword{internal}
