% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paintr-utils.R
\name{expr_pluck<-}
\alias{expr_pluck<-}
\title{Replace an Expression by Index Path}
\usage{
expr_pluck(.x, index_path) <- value
}
\arguments{
\item{.x}{An expression-like object.}

\item{index_path}{An index vector.}

\item{value}{A replacement value.}
}
\value{
The updated expression object.
}
\description{
Replace an Expression by Index Path
}
\keyword{internal}
