% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paintr-utils.R
\name{expr_pluck}
\alias{expr_pluck}
\title{Pluck an Expression by Index Path}
\usage{
expr_pluck(.x, index_path)
}
\arguments{
\item{.x}{An expression-like object.}

\item{index_path}{An index vector.}
}
\value{
The plucked expression or \code{NULL}.
}
\description{
Pluck an Expression by Index Path
}
\keyword{internal}
