% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paintr-utils.R
\name{expr_remove_emptycall2}
\alias{expr_remove_emptycall2}
\title{Remove Empty Non-ggplot Calls}
\usage{
expr_remove_emptycall2(.expr)
}
\arguments{
\item{.expr}{An expression object.}
}
\value{
The cleaned expression or \code{NULL}.
}
\description{
Remove Empty Non-ggplot Calls
}
\keyword{internal}
