% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paintr-utils.R
\name{expr_remove_null}
\alias{expr_remove_null}
\title{Remove Placeholder Marker Symbols}
\usage{
expr_remove_null(
  .expr,
  target = rlang::sym("_NULL_PLACEHOLDER"),
  current_path = numeric()
)
}
\arguments{
\item{.expr}{An expression object.}

\item{target}{The placeholder marker symbol.}

\item{current_path}{Internal recursion path.}
}
\value{
The cleaned expression.
}
\description{
Remove Placeholder Marker Symbols
}
\keyword{internal}
