% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paintr-runtime.R
\name{expr_replace_keywords}
\alias{expr_replace_keywords}
\title{Replace One Placeholder in an Expression}
\usage{
expr_replace_keywords(.expr, keyword, index_path, input_item)
}
\arguments{
\item{.expr}{An expression object.}

\item{keyword}{A placeholder symbol.}

\item{index_path}{An index path.}

\item{input_item}{A user input value.}
}
\value{
The updated expression.
}
\description{
Replace One Placeholder in an Expression
}
\keyword{internal}
