% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paintr-utils.R
\name{expr_type}
\alias{expr_type}
\title{Determine an Expression's Structural Type}
\usage{
expr_type(x)
}
\arguments{
\item{x}{An R expression object.}
}
\value{
A single string describing the expression type.
}
\description{
Internal helper used while traversing ggplot-like calls.
}
\keyword{internal}
