% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{flipHandler}
\alias{flipHandler}
\title{Module server for \code{coord_flip()}}
\usage{
flipHandler(id, module_id)
}
\arguments{
\item{id}{a string which becomes a name space for the module server}

\item{module_id}{the id of the ui element (not including its prefix created by
the name space) which determines whether or not to call \code{coord_flip()}}
}
\value{
\code{NULL} or \code{coord_flip()}
}
\description{
Module server for \code{coord_flip()}
}
\examples{
\dontrun{
flipHandler("my_boxplot", "settingFlip")
}
}
