% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paintr-ui.R
\name{generate_ui_expr}
\alias{generate_ui_expr}
\title{Build Expression Placeholder UI}
\usage{
generate_ui_expr(id, param)
}
\arguments{
\item{id}{Placeholder id.}

\item{param}{Parameter label.}
}
\value{
A \code{textInput()} UI object.
}
\description{
Build Expression Placeholder UI
}
\keyword{internal}
