% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paintr-ui.R
\name{generate_ui_num}
\alias{generate_ui_num}
\title{Build Numeric Placeholder UI}
\usage{
generate_ui_num(id, param)
}
\arguments{
\item{id}{Placeholder id.}

\item{param}{Parameter label.}
}
\value{
A \code{numericInput()} UI object.
}
\description{
Build Numeric Placeholder UI
}
\keyword{internal}
