% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paintr-ui.R
\name{generate_ui_text}
\alias{generate_ui_text}
\title{Build Text Placeholder UI}
\usage{
generate_ui_text(id, param)
}
\arguments{
\item{id}{Placeholder id.}

\item{param}{Parameter label.}
}
\value{
A \code{textInput()} UI object.
}
\description{
Build Text Placeholder UI
}
\keyword{internal}
