% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paintr-ui.R
\name{generate_ui_upload}
\alias{generate_ui_upload}
\title{Build Upload UI for a Placeholder}
\usage{
generate_ui_upload(id, copy_rules = NULL)
}
\arguments{
\item{id}{Placeholder id.}

\item{copy_rules}{Effective or user-supplied copy rules.}
}
\value{
A Shiny UI object.
}
\description{
Build Upload UI for a Placeholder
}
\keyword{internal}
