% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paintr-ui.R
\name{generate_ui_var}
\alias{generate_ui_var}
\title{Build a Variable Selector UI}
\usage{
generate_ui_var(data_var, id, param)
}
\arguments{
\item{data_var}{Character vector of candidate variables.}

\item{id}{Placeholder id.}

\item{param}{Parameter label.}
}
\value{
A \code{pickerInput()} UI object or \code{NULL}.
}
\description{
Build a Variable Selector UI
}
\keyword{internal}
