% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paintr-ui.R
\name{generate_ui_var_placeholder}
\alias{generate_ui_var_placeholder}
\title{Build a Deferred \code{var} UI Placeholder}
\usage{
generate_ui_var_placeholder(id)
}
\arguments{
\item{id}{Placeholder id.}
}
\value{
A placeholder \code{uiOutput()}.
}
\description{
Build a Deferred \code{var} UI Placeholder
}
\keyword{internal}
