% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paintr-utils.R
\name{get_expr_param}
\alias{get_expr_param}
\title{Read a Parameter Name from an Expression Path}
\usage{
get_expr_param(.expr, .path)
}
\arguments{
\item{.expr}{A call expression.}

\item{.path}{An index path.}
}
\value{
The parameter name or \code{NULL}.
}
\description{
Read a Parameter Name from an Expression Path
}
\keyword{internal}
