% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paintr.R
\name{get_filter}
\alias{get_filter}
\title{Extract filtering code}
\usage{
get_filter(data, search)
}
\arguments{
\item{data}{dataset}

\item{search}{filter conditions}
}
\value{
filtering code
}
\description{
Extract filtering code
}
