% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paintr-utils.R
\name{get_fun_names}
\alias{get_fun_names}
\title{Get a Call or Symbol Name}
\usage{
get_fun_names(x)
}
\arguments{
\item{x}{An expression object.}
}
\value{
A function or symbol name when available.
}
\description{
Get a Call or Symbol Name
}
\keyword{internal}
