% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paintr-runtime.R
\name{get_parameter_name_by_index_path}
\alias{get_parameter_name_by_index_path}
\title{Get the Argument Name for an Index Path}
\usage{
get_parameter_name_by_index_path(.expr, index_path)
}
\arguments{
\item{.expr}{An expression object.}

\item{index_path}{An index path.}
}
\value{
The argument name or \code{NULL}.
}
\description{
Get the Argument Name for an Index Path
}
\keyword{internal}
