% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geomGenerator.R
\name{ggGeomGenerator}
\alias{ggGeomGenerator}
\title{Title}
\usage{
ggGeomGenerator(
  id,
  data,
  geom_FUN,
  id_list,
  params_list,
  color_fill = FALSE,
  color_group = FALSE,
  userFUN = NULL,
  ...
)
}
\arguments{
\item{...}{arguments that go into \code{userFUN}}

\item{userFun}{a function that returns a named list, where the names of
this named list are parameters (except for \code{mapping}) of \code{geom_FUN}, and the elements
of this list are arguments of the corresponding parameters}
}
\value{

}
\description{
Title
}
