% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paintr_server_handler.R
\name{ggGeomHandler}
\alias{ggGeomHandler}
\title{Generate the geom component of a \code{ggplot2} plot}
\usage{
ggGeomHandler(
  id,
  data,
  geom_FUN,
  id_list,
  params_list,
  color_fill = FALSE,
  color_group = FALSE,
  userFUN = NULL,
  ...
)
}
\arguments{
\item{id}{An ID string that corresponds with the ID used to call the module's UI function.}

\item{data}{data}

\item{geom_FUN}{\verb{geom_<chart>}}

\item{id_list}{list of id of all ui elements}

\item{params_list}{list of parameters that the ui elements should correspond to}

\item{color_fill}{bool; optional. Whether or not to use the same variable for both color and fill}

\item{color_group}{bool; optional. Whether or not to use the same variable for both color and group}

\item{userFUN}{a function that returns a named list, where the names of
this named list are parameters (except for \code{mapping}) of \verb{geom_<chart>}, and the elements
of this list are arguments of the corresponding parameters}

\item{...}{arguments that go into \code{userFUN}}
}
\value{
list of plot and code of a geom component of a \code{ggplot2} plot
}
\description{
Generate the geom component of a \code{ggplot2} plot
}
