% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggpaintr-package.R
\docType{package}
\name{ggpaintr-package}
\alias{ggpaintr}
\alias{ggpaintr-package}
\title{ggpaintr: Formula-Driven ggplot Shiny Apps}
\description{
\code{ggpaintr} provides the \code{ggpaintr} workflow for turning a single
ggplot-like formula string into a small Shiny app with generated controls,
rendered plots, generated code, upload support, and inline runtime feedback.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://willju-wangqian.github.io/ggpaintr/}
  \item \url{https://github.com/willju-wangqian/ggpaintr}
  \item Report bugs at \url{https://github.com/willju-wangqian/ggpaintr/issues}
}

}
\author{
\strong{Maintainer}: Wangqian Ju \email{wju@iastate.edu} (\href{https://orcid.org/0000-0002-9977-377X}{ORCID})

Authors:
\itemize{
  \item Jinji Pang \email{pjj0702@iastate.edu} (\href{https://orcid.org/0000-0002-2267-2313}{ORCID})
  \item Zhili Qiao \email{zlqiao@iastate.edu} (\href{https://orcid.org/0000-0002-8154-0898}{ORCID})
}

}
\keyword{internal}
