% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paintr-app.R
\name{ggpaintr_bind_draw}
\alias{ggpaintr_bind_draw}
\title{Bind Draw Behavior into a Shiny App}
\usage{
ggpaintr_bind_draw(input, ggpaintr_state, ids = ggpaintr_state$ids)
}
\arguments{
\item{input}{A Shiny \code{input} object.}

\item{ggpaintr_state}{A \code{ggpaintr_state} object.}

\item{ids}{A \code{ggpaintr_ids} object describing the top-level Shiny ids used by
the integration helpers.}
}
\value{
Invisibly returns \code{ggpaintr_state}.
}
\description{
Bind Draw Behavior into a Shiny App
}
\examples{
\dontrun{
server <- function(input, output, session) {
  ps <- ggpaintr_server_state("ggplot(mtcars, aes(x = var)) + geom_histogram()")
  ggpaintr_bind_control_panel(input, output, ps)
  ggpaintr_bind_draw(input, ps)
  ggpaintr_bind_plot(output, ps)
}
}
}
