% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paintr-app.R
\name{ggpaintr_error_value}
\alias{ggpaintr_error_value}
\title{Return Default Error UI from a Runtime Result}
\usage{
ggpaintr_error_value(runtime_result)
}
\arguments{
\item{runtime_result}{A runtime result list returned by \code{paintr_build_runtime()}.}
}
\value{
A Shiny tag or \code{NULL}.
}
\description{
Return Default Error UI from a Runtime Result
}
