% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paintr-utils.R
\name{handle_call_break_sum}
\alias{handle_call_break_sum}
\title{Flatten \code{+}-Separated Plot Layers}
\usage{
handle_call_break_sum(x)
}
\arguments{
\item{x}{An expression or call.}
}
\value{
A nested structure with plot layers separated.
}
\description{
Flatten \code{+}-Separated Plot Layers
}
\keyword{internal}
