% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paintr-utils.R
\name{handle_duplicate_names}
\alias{handle_duplicate_names}
\title{Suffix Duplicate Layer Names}
\usage{
handle_duplicate_names(x)
}
\arguments{
\item{x}{A character vector of names.}
}
\value{
A character vector with duplicates made unique.
}
\description{
Suffix Duplicate Layer Names
}
\keyword{internal}
