% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paintr-runtime.R
\name{handle_num}
\alias{handle_num}
\title{Replace a \code{num} Placeholder}
\usage{
handle_num(.expr, index_path, input_item)
}
\arguments{
\item{.expr}{An expression object.}

\item{index_path}{An index path.}

\item{input_item}{A user input value.}
}
\value{
The updated expression.
}
\description{
Replace a \code{num} Placeholder
}
\keyword{internal}
