% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paintr-runtime.R
\name{handle_unknown}
\alias{handle_unknown}
\title{Leave Unknown Keywords Untouched}
\usage{
handle_unknown(.expr, keyword)
}
\arguments{
\item{.expr}{An expression object.}

\item{keyword}{A keyword expression.}
}
\value{
The original expression.
}
\description{
Leave Unknown Keywords Untouched
}
\keyword{internal}
