% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paintr-runtime.R
\name{handle_upload}
\alias{handle_upload}
\title{Replace an \code{upload} Placeholder}
\usage{
handle_upload(.expr, index_path, input_item)
}
\arguments{
\item{.expr}{An expression object.}

\item{index_path}{An index path.}

\item{input_item}{A user input value.}
}
\value{
The updated expression.
}
\description{
Replace an \code{upload} Placeholder
}
\keyword{internal}
