% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paintr_server_handler.R
\name{labsHandler}
\alias{labsHandler}
\title{Module server for \code{labs()}}
\usage{
labsHandler(id, module_id, param)
}
\arguments{
\item{id}{An ID string that corresponds with the ID used to call the module's UI function.}

\item{module_id}{the id of the ui elements (not including its prefix created by
the name space) which gives input as arguments of \code{labs()}}

\item{param}{parameters of \code{labs()} that correspond to the ui elements}
}
\value{
\code{labs()} and its code or \code{NULL}
}
\description{
Module server for \code{labs()}
}
\note{
the order of \code{param} should match the order of \code{module_id}. Currently the
following parameters of \code{labs()} are implemented
\itemize{
\item \code{x}
\item \code{y}
\item \code{title}
\item \code{subtitle}
}
}
