% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paintr_ui.R
\name{mappingGroupUI}
\alias{mappingGroupUI}
\title{UI component and its id for aesthetic mapping of group}
\usage{
mappingGroupUI(id, data_vars, ui_id = "mapGroup")
}
\arguments{
\item{id}{An ID string that corresponds with the ID used to call the module's UI function.}

\item{data_vars}{variable names of the dataset}

\item{ui_id}{id of the ui element}
}
\value{
a list with two elements
\itemize{
\item ui: the ui element of this piece
\item id: the id of this ui element
}
}
\description{
UI component and its id for aesthetic mapping of group
}
