% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paintr_ui.R
\name{multipleColorPickerUI}
\alias{multipleColorPickerUI}
\title{ui component for setting \code{colourInput}}
\usage{
multipleColorPickerUI(id, init_colors, labels, ui_id = "colorPicker")
}
\arguments{
\item{id}{An ID string that corresponds with the ID used to call the module's UI function.}

\item{init_colors}{initial colors for \code{colourInput}}

\item{labels}{labels for \code{colourInput}}

\item{ui_id}{id of the ui element}
}
\value{
a list with two elements
\itemize{
\item ui: the ui element of this piece
\item id: the id of this ui element
}
}
\description{
ui component for setting \code{colourInput}
}
