% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paintr-ui.R
\name{paintr_attach_help}
\alias{paintr_attach_help}
\title{Attach Optional Help Text to a UI Control}
\usage{
paintr_attach_help(ui, help = NULL)
}
\arguments{
\item{ui}{A Shiny UI object.}

\item{help}{Optional help text.}
}
\value{
A Shiny UI object.
}
\description{
Attach Optional Help Text to a UI Control
}
\keyword{internal}
