% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paintr-app.R
\name{paintr_build_app_ui}
\alias{paintr_build_app_ui}
\title{Build the Standard ggpaintr App UI}
\usage{
paintr_build_app_ui(title_label, draw_label, export_label)
}
\arguments{
\item{title_label}{App title text.}

\item{draw_label}{Draw button text.}

\item{export_label}{Export button text.}
}
\value{
A Shiny UI object.
}
\description{
Build the Standard ggpaintr App UI
}
\keyword{internal}
