% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paintr-ui.R
\name{paintr_build_ui_list}
\alias{paintr_build_ui_list}
\title{Build a UI List for a Parsed Formula}
\usage{
paintr_build_ui_list(paintr_obj, copy_rules = NULL)
}
\arguments{
\item{paintr_obj}{A \code{paintr_obj}.}

\item{copy_rules}{Effective or user-supplied copy rules.}
}
\value{
A named list of UI controls by layer.
}
\description{
Build a UI List for a Parsed Formula
}
\keyword{internal}
