% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paintr-copy.R
\name{paintr_compact_copy_rule_branch}
\alias{paintr_compact_copy_rule_branch}
\title{Recursively Compact Copy Rules Against Defaults}
\usage{
paintr_compact_copy_rule_branch(current, defaults = NULL)
}
\arguments{
\item{current}{A current copy-rule branch.}

\item{defaults}{A default copy-rule branch.}
}
\value{
A compacted branch or \code{NULL}.
}
\description{
Recursively Compact Copy Rules Against Defaults
}
\keyword{internal}
