% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paintr-runtime.R
\name{paintr_complete_expr}
\alias{paintr_complete_expr}
\title{Complete a Parsed Formula with User Inputs}
\usage{
paintr_complete_expr(paintr_obj, input, envir = parent.frame())
}
\arguments{
\item{paintr_obj}{A \code{paintr_obj}.}

\item{input}{A Shiny input-like object.}

\item{envir}{The environment used to resolve local data objects.}
}
\value{
A named list with \code{complete_expr_list}, \code{code_text}, and \code{eval_env}.
}
\description{
Complete a Parsed Formula with User Inputs
}
\keyword{internal}
