% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paintr-copy.R
\name{paintr_copy_leaf_fields}
\alias{paintr_copy_leaf_fields}
\title{Return Allowed Copy Leaf Fields}
\usage{
paintr_copy_leaf_fields()
}
\value{
A character vector.
}
\description{
Return Allowed Copy Leaf Fields
}
\keyword{internal}
