% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paintr-copy.R
\name{paintr_default_copy_rules}
\alias{paintr_default_copy_rules}
\title{Default Copy Rules for ggpaintr}
\usage{
paintr_default_copy_rules()
}
\value{
A named list of default copy rules.
}
\description{
Internal registry for all user-facing control copy. Maintainers should update
this object when they want to refine defaults or add support for new common
parameters in one place.
}
\keyword{internal}
