% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paintr-runtime.R
\name{paintr_error_ui}
\alias{paintr_error_ui}
\title{Build Inline Error UI for a Paintr App}
\usage{
paintr_error_ui(message)
}
\arguments{
\item{message}{A runtime error message.}
}
\value{
A Shiny tag or \code{NULL}.
}
\description{
Build Inline Error UI for a Paintr App
}
\keyword{internal}
