% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paintr-export.R
\name{paintr_escape_string_segment}
\alias{paintr_escape_string_segment}
\title{Escape One String Segment for a Template Literal}
\usage{
paintr_escape_string_segment(x)
}
\arguments{
\item{x}{A character string.}
}
\value{
A single escaped string segment.
}
\description{
Escape One String Segment for a Template Literal
}
\keyword{internal}
