% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paintr.R
\name{paintr_geom_construct}
\alias{paintr_geom_construct}
\title{Deconstruct the paintr expression}
\usage{
paintr_geom_construct(expr)
}
\arguments{
\item{expr}{a \code{ggplot2} alike expression}
}
\value{
a list of pieces of the keywords
}
\description{
Deconstruct the paintr expression
}
