% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paintr-runtime.R
\name{paintr_get_plot_safe}
\alias{paintr_get_plot_safe}
\title{Safely Build a Plot from a Runtime Result}
\usage{
paintr_get_plot_safe(runtime_result, envir = parent.frame())
}
\arguments{
\item{runtime_result}{A runtime result list.}

\item{envir}{A fallback evaluation environment.}
}
\value{
An updated runtime result.
}
\description{
Safely Build a Plot from a Runtime Result
}
\keyword{internal}
