% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paintr-ui.R
\name{paintr_get_tab_ui}
\alias{paintr_get_tab_ui}
\title{Build the Tabbed Control UI for a Parsed Formula}
\usage{
paintr_get_tab_ui(paintr_obj)
}
\arguments{
\item{paintr_obj}{A \code{paintr_obj}.}
}
\value{
A Shiny UI object or \code{NULL}.
}
\description{
Build the Tabbed Control UI for a Parsed Formula
}
\keyword{internal}
