% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paintr.R
\name{paintr_get_ui}
\alias{paintr_get_ui}
\title{Get a ui element or its id of a paintr keyword from a \code{paintr_obj}}
\usage{
paintr_get_ui(
  paintr_obj,
  selected_ui_name,
  type = "ui",
  scope = NULL,
  verbose = FALSE
)
}
\arguments{
\item{paintr_obj}{a \code{paintr_obj}}

\item{selected_ui_name}{the keyword of the desired ui element}

\item{type}{optional. \code{type} can be \code{ui} or \code{id}}

\item{scope}{one value of \code{mapping}, \code{geom_args}, or \code{plot_settings}. Used to
distinguish keywords with the same name but in different scope. For example \code{size}
can be either a \code{mapping} keyword or a \code{geom_args} keyword.}

\item{verbose}{Whether or not to send warning messages when ui element is not found}
}
\value{
the ui or id of \code{selected_ui_name}
}
\description{
Get a ui element or its id of a paintr keyword from a \code{paintr_obj}
}
\examples{
ptr_obj <- paintr("boxplot_id", mtcars, geom_boxplot(aes(x, y)))
paintr_get_ui(ptr_obj, "x")
}
