% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paintr-upload.R
\name{paintr_get_uploaded_data}
\alias{paintr_get_uploaded_data}
\title{Return Uploaded Data Only}
\usage{
paintr_get_uploaded_data(input, upload_id)
}
\arguments{
\item{input}{A Shiny input-like object.}

\item{upload_id}{The upload placeholder id.}
}
\value{
The uploaded object or \code{NULL}.
}
\description{
Return Uploaded Data Only
}
\keyword{internal}
