% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paintr-copy.R
\name{paintr_interpolate_copy_text}
\alias{paintr_interpolate_copy_text}
\title{Replace Copy Template Tokens}
\usage{
paintr_interpolate_copy_text(text, param = NULL, layer_name = NULL)
}
\arguments{
\item{text}{A text template.}

\item{param}{A parameter label.}

\item{layer_name}{A layer name.}
}
\value{
A single string or \code{NULL}.
}
\description{
Replace Copy Template Tokens
}
\keyword{internal}
