% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paintr-copy.R
\name{paintr_merge_copy_rules}
\alias{paintr_merge_copy_rules}
\title{Recursively Merge Copy Rules}
\usage{
paintr_merge_copy_rules(base, overrides)
}
\arguments{
\item{base}{A base list.}

\item{overrides}{Override values.}
}
\value{
A merged list.
}
\description{
Recursively Merge Copy Rules
}
\keyword{internal}
