% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paintr-data.R
\name{paintr_normalize_column_name_vector}
\alias{paintr_normalize_column_name_vector}
\title{Normalize a Vector of Column Names}
\usage{
paintr_normalize_column_name_vector(x, n = length(x))
}
\arguments{
\item{x}{A character vector of names.}

\item{n}{Expected column count.}
}
\value{
A normalized character vector with syntactic, unique names.
}
\description{
Normalize a Vector of Column Names
}
\keyword{internal}
