% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paintr-copy.R
\name{paintr_normalize_copy_rules}
\alias{paintr_normalize_copy_rules}
\title{Normalize User Copy Rule Keys}
\usage{
paintr_normalize_copy_rules(copy_rules)
}
\arguments{
\item{copy_rules}{User-supplied copy rules.}
}
\value{
A normalized copy-rule list.
}
\description{
Normalize User Copy Rule Keys
}
\keyword{internal}
