% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paintr-copy.R
\name{paintr_normalize_param_key}
\alias{paintr_normalize_param_key}
\title{Normalize a Copy Rule Parameter Key}
\usage{
paintr_normalize_param_key(param)
}
\arguments{
\item{param}{A parameter name or \code{NULL}.}
}
\value{
A normalized parameter key.
}
\description{
Normalize a Copy Rule Parameter Key
}
\keyword{internal}
