% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paintr-data.R
\name{paintr_normalize_tabular_data}
\alias{paintr_normalize_tabular_data}
\title{Coerce an Object to Tabular Data for \code{ggpaintr}}
\usage{
paintr_normalize_tabular_data(data, source = "Data")
}
\arguments{
\item{data}{An arbitrary R object.}

\item{source}{A short label for error messages.}
}
\value{
A \code{data.frame}-like object with normalized column names.
}
\description{
Coerce an Object to Tabular Data for \code{ggpaintr}
}
\keyword{internal}
