% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paintr-copy.R
\name{paintr_param_is_unnamed}
\alias{paintr_param_is_unnamed}
\title{Detect Whether a Parameter Is Unnamed}
\usage{
paintr_param_is_unnamed(param)
}
\arguments{
\item{param}{A parameter value from \code{paintr_obj$param_list}.}
}
\value{
A single logical value.
}
\description{
Detect Whether a Parameter Is Unnamed
}
\keyword{internal}
