% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paintr.R
\name{paintr_plot_code}
\alias{paintr_plot_code}
\title{Generate plot and the corresponding code of \code{ggplot2} from a \code{paintr_obj}}
\usage{
paintr_plot_code(
  paintr_obj,
  data_filter = "",
  selected_color_rctv = NULL,
  selected_fill_rctv = NULL,
  color_fill = FALSE,
  color_group = FALSE,
  userFUN = NULL,
  ...
)
}
\arguments{
\item{paintr_obj}{a \code{paintr_obj}}

\item{data_filter}{data filtering code}

\item{selected_color_rctv}{reactive value returned by \code{scaleColor_build_reactivity} for color}

\item{selected_fill_rctv}{reactive value returned by \code{scaleColor_build_reactivity} for fill}

\item{color_fill}{bool; optional. Whether or not to use the same variable for both color and fill}

\item{color_group}{bool; optional. Whether or not to use the same variable for both color and group}

\item{userFUN}{a function that returns a named list, where the names of
this named list are parameters (except for \code{mapping}) of \verb{geom_<chart>}, and the elements
of this list are arguments of the corresponding parameters}

\item{...}{arguments that go into \code{userFUN}}
}
\value{
a named list of two elements; plot and code
}
\description{
Generate plot and the corresponding code of \code{ggplot2} from a \code{paintr_obj}
}
\note{
this function should be called inside \code{observeEvent()} since the \code{isolate()} scope provided
by \code{observeEvent()} is essential.
}
