% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paintr-export.R
\name{paintr_prefix_assignment}
\alias{paintr_prefix_assignment}
\title{Prefix Serialized Template Lines with an Assignment}
\usage{
paintr_prefix_assignment(name, value_lines)
}
\arguments{
\item{name}{An assignment target.}

\item{value_lines}{A character vector of serialized value lines.}
}
\value{
A character vector.
}
\description{
Prefix Serialized Template Lines with an Assignment
}
\keyword{internal}
