% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paintr-upload.R
\name{paintr_prepare_eval_env}
\alias{paintr_prepare_eval_env}
\title{Clone an Evaluation Environment and Inject Uploads}
\usage{
paintr_prepare_eval_env(paintr_obj, input, envir = parent.frame())
}
\arguments{
\item{paintr_obj}{A \code{paintr_obj}.}

\item{input}{A Shiny input-like object.}

\item{envir}{A parent environment.}
}
\value{
An evaluation environment containing uploaded datasets and any other
placeholder-provided objects.
}
\description{
Clone an Evaluation Environment and Inject Uploads
}
\keyword{internal}
