% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paintr-upload.R
\name{paintr_read_uploaded_data}
\alias{paintr_read_uploaded_data}
\title{Read Uploaded Paintr Data}
\usage{
paintr_read_uploaded_data(file_info)
}
\arguments{
\item{file_info}{A Shiny \code{fileInput()} value.}
}
\value{
The uploaded object or \code{NULL}.
}
\description{
Read Uploaded Paintr Data
}
\keyword{internal}
