% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paintr-data.R
\name{paintr_reserved_words}
\alias{paintr_reserved_words}
\title{Reserved Words That Should Not Be Used as Column Names}
\usage{
paintr_reserved_words()
}
\value{
A character vector of reserved words.
}
\description{
Reserved Words That Should Not Be Used as Column Names
}
\keyword{internal}
