% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paintr-export.R
\name{paintr_serialize_r_object}
\alias{paintr_serialize_r_object}
\title{Serialize an R Object as Source Text}
\usage{
paintr_serialize_r_object(x)
}
\arguments{
\item{x}{An R object.}
}
\value{
A character vector.
}
\description{
Serialize an R Object as Source Text
}
\keyword{internal}
