% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paintr-upload.R
\name{paintr_upload_default_name}
\alias{paintr_upload_default_name}
\title{Derive a Default Object Name from an Uploaded File}
\usage{
paintr_upload_default_name(file_name)
}
\arguments{
\item{file_name}{The uploaded filename.}
}
\value{
A syntactic R object name.
}
\description{
Derive a Default Object Name from an Uploaded File
}
\keyword{internal}
