% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paintr-runtime.R
\name{paintr_validate_plot_render_safe}
\alias{paintr_validate_plot_render_safe}
\title{Safely Validate a Plot at Render Time}
\usage{
paintr_validate_plot_render_safe(runtime_result)
}
\arguments{
\item{runtime_result}{A runtime result list.}
}
\value{
An updated runtime result.
}
\description{
Safely Validate a Plot at Render Time
}
\keyword{internal}
