% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paintr-app.R
\name{ptr_extract_code}
\alias{ptr_extract_code}
\title{Return Generated Code Text from a Runtime Result}
\usage{
ptr_extract_code(runtime_result)
}
\arguments{
\item{runtime_result}{A runtime result list returned by \code{ptr_exec()}.}
}
\value{
A character string or \code{NULL}.
}
\description{
Return Generated Code Text from a Runtime Result
}
