% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paintr-app.R
\name{ptr_extract_error}
\alias{ptr_extract_error}
\title{Return Default Error UI from a Runtime Result}
\usage{
ptr_extract_error(runtime_result)
}
\arguments{
\item{runtime_result}{A runtime result list returned by \code{ptr_exec()}.}
}
\value{
A Shiny tag or \code{NULL}.
}
\description{
Return Default Error UI from a Runtime Result
}
