% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paintr-server.R
\name{ptr_register}
\alias{ptr_register}
\alias{ptr_register_plot}
\alias{ptr_register_error}
\alias{ptr_register_code}
\title{Register Plot / Error / Code Outputs}
\usage{
ptr_register_plot(output, state)

ptr_register_error(output, state)

ptr_register_code(output, state)
}
\arguments{
\item{output}{The Shiny output object.}

\item{state}{A \code{ptr_state} from \code{\link[=ptr_init_state]{ptr_init_state()}}.}
}
\value{
The \code{output} object, invisibly.
}
\description{
Attaches the standard plot, error, and code outputs to the Shiny \code{output}
object. Already invoked by \code{\link[=ptr_server]{ptr_server()}}; exposed for embedders that
compose outputs manually.
}
