% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paintr-app.R
\name{ptr_register_code}
\alias{ptr_register_code}
\title{Bind Default Code Rendering into a Shiny App}
\usage{
ptr_register_code(output, ptr_state)
}
\arguments{
\item{output}{A Shiny \code{output} object.}

\item{ptr_state}{A \code{ptr_state} object.}
}
\value{
Invisibly returns \code{ptr_state}.
}
\description{
Bind Default Code Rendering into a Shiny App
}
\examples{
if (interactive()) {
server <- function(input, output, session) {
  ps <- ptr_server_state("ggplot(mtcars, aes(x = var)) + geom_histogram()")
  ptr_register_draw(input, ps)
  ptr_register_code(output, ps)
}
}
}
