% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paintr-app.R
\name{ptr_register_draw}
\alias{ptr_register_draw}
\title{Bind Draw Behavior into a Shiny App}
\usage{
ptr_register_draw(input, ptr_state)
}
\arguments{
\item{input}{A Shiny \code{input} object.}

\item{ptr_state}{A \code{ptr_state} object.}
}
\value{
Invisibly returns \code{ptr_state}.
}
\description{
Bind Draw Behavior into a Shiny App
}
\examples{
if (interactive()) {
server <- function(input, output, session) {
  ps <- ptr_server_state("ggplot(mtcars, aes(x = var)) + geom_histogram()")
  ptr_setup_controls(input, output, ps)
  ptr_register_draw(input, ps)
  ptr_register_plot(output, ps)
}
}
}
