% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paintr_control.R
\name{scaleColor_build_reactivity}
\alias{scaleColor_build_reactivity}
\title{Build reactivity for \code{scale_color} or \code{scale_fill}}
\usage{
scaleColor_build_reactivity(id, paintr_rctv, color_or_fill)
}
\arguments{
\item{id}{An ID string that corresponds with the ID used for all component of this \code{paintr_obj}}

\item{paintr_rctv}{a reactive value of a \code{paintr_obj} object created by \code{paintr()}}

\item{color_or_fill}{string. either \code{color} or \code{fill}. Specifies whether it's \code{scale_color}
or \code{scale_fill}}
}
\value{
a reactive object that is passed into \code{paintr_plot_code()}
}
\description{
Build reactivity for \code{scale_color} or \code{scale_fill}
}
