% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moreFuncs.R
\name{scaleColorWrapper}
\alias{scaleColorWrapper}
\title{Title}
\usage{
scaleColorWrapper(id, paintr_obj, data, color_name, scaleColor_name)
}
\arguments{
\item{scaleColor_name}{}
}
\value{

}
\description{
Title
}
