% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paintr_server_handler.R
\name{stringParamHandler}
\alias{stringParamHandler}
\title{Module server for a \code{ggplot2} function which takes strings as arguments}
\usage{
stringParamHandler(id, module_id, param, FUN)
}
\arguments{
\item{id}{An ID string that corresponds with the ID used to call the module's UI function.}

\item{module_id}{the id of the ui elements (not including its prefix created by
the name space) which gives input as arguments of \code{FUN()}}

\item{param}{parameters of \code{theme()} that correspond to the ui elements}

\item{FUN}{a \code{ggplot2} function which takes strings as arguments}
}
\value{
the return of \code{FUN} and its code or \code{NULL}
}
\description{
Module server for a \code{ggplot2} function which takes strings as arguments
}
