% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paintr-ui.R
\name{tab_wrap_ui}
\alias{tab_wrap_ui}
\title{Wrap Layer Controls in Tabs}
\usage{
tab_wrap_ui(ui_list)
}
\arguments{
\item{ui_list}{A named list of UI lists.}
}
\value{
A \code{tabsetPanel()}.
}
\description{
Wrap Layer Controls in Tabs
}
\keyword{internal}
