% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{themeCollector}
\alias{themeCollector}
\title{collecting all arguments for theme()}
\usage{
themeCollector(input)
}
\arguments{
\item{input}{}
}
\value{

}
\description{
collecting all arguments for theme()
}
