% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paintr_server_handler.R
\name{themeHandler}
\alias{themeHandler}
\title{Module server for \code{theme()}}
\usage{
themeHandler(id, module_id, param)
}
\arguments{
\item{id}{An ID string that corresponds with the ID used to call the module's UI function.}

\item{module_id}{the id of the ui elements (not including its prefix created by
the name space) which gives input as arguments of \code{theme()}}

\item{param}{parameters of \code{theme()} that correspond to the ui elements}
}
\value{
\code{theme()} and its code or \code{NULL}
}
\description{
Module server for \code{theme()}
}
\note{
the order of \code{param} should match the order of \code{module_id}. Currently the
following parameters of \code{theme()} are implemented
\itemize{
\item \code{legend.position}
\item \code{legend.direction}
\item \code{legend.box}
}
}
