% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paintr_control.R
\name{connect_param_id}
\alias{connect_param_id}
\title{Connect parameters to their input obtained from the ui elements by their ids}
\usage{
connect_param_id(
  session_input,
  id_list,
  params,
  color_fill = FALSE,
  color_group = FALSE
)
}
\arguments{
\item{session_input}{input of a shiny session}

\item{id_list}{list of ids}

\item{params}{list of parameters}

\item{color_fill}{bool; optional. Whether or not to use the same variable for both color and fill}

\item{color_group}{bool; optional. Whether or not to use the same variable for both color and group}
}
\value{
a named list which has parameters as names and ids as list elements
}
\description{
Connect parameters to their input obtained from the ui elements by their ids
}
