% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uiControls.R
\name{controlUI}
\alias{controlUI}
\title{Title}
\usage{
controlUI(
  id,
  data_vars,
  mapping,
  defaultArgs,
  geom_args = NULL,
  plot_settings = NULL,
  extra_uiFunc = NULL,
  extra_uiFuncArgs = NULL
)
}
\arguments{
\item{extra_uiFuncArgs}{}
}
\value{

}
\description{
Title
}
