% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paintr_server_handler.R
\name{flipHandler}
\alias{flipHandler}
\title{Module server for \code{coord_flip()}}
\usage{
flipHandler(id, module_id)
}
\arguments{
\item{id}{An ID string that corresponds with the ID used to call the module's UI function.}

\item{module_id}{the id of the ui element (not including its prefix created by
the name space) which determines whether or not to call \code{coord_flip()}}
}
\value{
\code{coord_flip()} and its code or \code{NULL}
}
\description{
Module server for \code{coord_flip()}
}
