% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paintr-ui.R
\name{generate_ui_individual}
\alias{generate_ui_individual}
\title{Build UI for One Placeholder}
\usage{
generate_ui_individual(key, id, param)
}
\arguments{
\item{key}{Placeholder keyword.}

\item{id}{Placeholder id.}

\item{param}{Parameter label.}
}
\value{
A Shiny UI object or \code{NULL}.
}
\description{
Build UI for One Placeholder
}
\keyword{internal}
