% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paintr-ui.R
\name{generate_ui_upload}
\alias{generate_ui_upload}
\title{Build Upload UI for a Placeholder}
\usage{
generate_ui_upload(id)
}
\arguments{
\item{id}{Placeholder id.}
}
\value{
A Shiny UI object with an attached \code{ui_expr} attribute.
}
\description{
Build Upload UI for a Placeholder
}
\keyword{internal}
