% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paintr-app.R
\name{paintr_app_components}
\alias{paintr_app_components}
\title{Build Reusable App Components for ggpaintr}
\usage{
paintr_app_components(formula, envir = parent.frame())
}
\arguments{
\item{formula}{A single formula string.}

\item{envir}{Environment used to resolve local data objects.}
}
\value{
A list with \code{ui} and \code{server}.
}
\description{
Build Reusable App Components for ggpaintr
}
\keyword{internal}
