% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paintr-copy.R
\name{paintr_copy_keywords}
\alias{paintr_copy_keywords}
\title{Return Allowed Copy Keywords}
\usage{
paintr_copy_keywords()
}
\value{
A character vector.
}
\description{
Return Allowed Copy Keywords
}
\keyword{internal}
