% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paintr-app.R
\name{ptr_register_code}
\alias{ptr_register_code}
\title{Bind Default Code Rendering into a Shiny App}
\usage{
ptr_register_code(output, ptr_state, ids = ptr_state$ids)
}
\arguments{
\item{output}{A Shiny \code{output} object.}

\item{ptr_state}{A \code{ptr_state} object.}

\item{ids}{A \code{ptr_build_ids} object describing the top-level Shiny ids used by
the integration helpers.}
}
\value{
Invisibly returns \code{ptr_state}.
}
\description{
Bind Default Code Rendering into a Shiny App
}
\examples{
\dontrun{
server <- function(input, output, session) {
  ps <- ptr_server_state("ggplot(mtcars, aes(x = var)) + geom_histogram()")
  ptr_register_draw(input, ps)
  ptr_register_code(output, ps)
}
}
}
