% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paintr-app.R
\name{ptr_setup_controls}
\alias{ptr_setup_controls}
\title{Bind the Generated Control Panel into a Shiny App}
\usage{
ptr_setup_controls(input, output, ptr_state)
}
\arguments{
\item{input}{A Shiny \code{input} object.}

\item{output}{A Shiny \code{output} object.}

\item{ptr_state}{A \code{ptr_state} object.}
}
\value{
Invisibly returns \code{ptr_state}.
}
\description{
Register the dynamic \code{var} controls and render the standard tabbed control
panel into the target \code{uiOutput()}.
}
\examples{
if (interactive()) {
server <- function(input, output, session) {
  ps <- ptr_server_state("ggplot(mtcars, aes(x = var)) + geom_histogram()")
  ptr_setup_controls(input, output, ps)
  ptr_register_draw(input, ps)
  ptr_register_plot(output, ps)
}
}
}
